Add step to build vscode lib in contributing doc
The top-level `pnpm build` will only build the `napi/*` packages, but the `editors/vscode` package needs to be built in order for the `editors/vscode/oxc_language_server.vsix` file referenced in the next steps to be generated. I'm not actually sure if the `napi/*` packages are even required to be built, so maybe step 2 could be removed as well.