oxlint JS plugins blog post (#542)
Draft blog post announcing Oxlint JS plugins.

It's not quite as polished as I'd like it to be, but we need to get it out, and I've run out of time, having prioritised getting the selectors DSL implemented.

The benchmarks are a bit weird, for 2 reasons:

1. In other rough benchmarks, we've seen JS plugins have a much higher cost. These benchmark numbers were originally from Cam's laptop, and I've replicated the result on my MacBook Air M3. But I ran into the same problem I had before that adding a JS plugin actually *improves* performance. This is clearly a nonsense - that's a side effect of either reduced file system contention, or some oddity of thread scheduling on M-series chips with efficiency cores. But getting to the bottom of that would take days, or maybe even weeks.

2. Oxlint is coming out at only 10x faster than ESLint. I think we'd expect a lot more than that. As a guess, it might be that ESLint in this case is not generating any errors, whereas Oxlint generates ~200 diagnostics. Printing diagnostics is really slow.

I actually don't think it matters that much if they're not super-scientific. Obviously, the performance is going to vary a lot depending on how many plugins, and how complex they are.

The benchmarks do still represent the thrust of the truth:

1. JS plugins are much faster than you'd expect.
2. Oxlint is way faster than ESLint, even with JS plugins in the mix.

In my opinion, I don't think it's going to be possible to get a more solid benchmark in a reasonable amount of time, so our choices are: either go with them or, if we're uncomfortable with how unscientific it is, just remove them.

The main point of the blog post is the feature, not the performance.