docs(website): Update instructions for adding rules to explain config documentation. (#578)
This is a bit wordy, but it's important that developers know how to properly document
their rules so that the website can generate it automatically and future tooling can rely on this behavior.

I'm unsure if there's a better way to generate the website than the command I note here. I tried using `just website` but I couldn't figure out a `path` value that would work.