docs(rules): Remove dead rules. (#615)
Most of these rules have been removed from the oxlint package for a long time.

I believe they were forgotten, and the existence of these pages is an accidental part of the way the website is generated.

I found all of these by deleting the entire `rules` dir and then regenerating the website.