chore: Update formatter documentation page (#667)
* Update formatter documentation page

Miscellaneous improvements to phrasing, and various miscellaneous clarifications to the doc page.

Signed-off-by: Connor Shea <connor.james.shea@gmail.com>

* Add a section about configuring oxfmt as a pre-commit hook.

This was something I ran into with switching to oxfmt in the oxlint-migrate repo.

And it looks like Boshen hit the same problem in oxc_walker: https://github.com/oxc-project/oxc-walker/commit/4d765c9d71dc2496846e2f43a6dd010184c74d2b

So it seems like it'd be a good idea to have this documented explicitly :)

* Rephrase the pre-commit hook section.

* Rephrase the pre-commit hook section again.

Correctly distinguish the `--no-error-on-unmatched-pattern` and `--ignore-unmatched` flags.

* Update src/docs/guide/usage/formatter.md

Signed-off-by: Connor Shea <connor.james.shea@gmail.com>

* Update src/docs/guide/usage/formatter.md

Co-authored-by: Yuji Sugiura <6259812+leaysgur@users.noreply.github.com>
Signed-off-by: Connor Shea <connor.james.shea@gmail.com>

---------

Signed-off-by: Connor Shea <connor.james.shea@gmail.com>
Co-authored-by: Yuji Sugiura <6259812+leaysgur@users.noreply.github.com>