docs(oxlint): update JS plugins API support (#683)
Update list of supported/unsupported features in docs for Oxlint JS plugins.

It had got out of date. Notably it still said that we don't support token-based APIs, which we now do.

I added custom file formats to the "unsupported" section.

I've also removed plugins-written-in-TS from the "supported" section, because it only *kind of* works - the types are a bit of a mess.